Refactored. If this works, follow on PR to move these checks into gofannon
